Arm works with both the joystick and the buttons.
The arm has been coded to work for buttons and joystick simultaneously, and the values are correctly calibrated. Using the gotoposition function, we can assign a set target position, and go to the position based on the current position (without having a set power direction); no longer need swing up and down separately.